requirements.txt
streamlit==1.28.0
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
numpy==1.24.3
plotly==5.17.0
pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
joblib==1.3.2
textblob==0.17.1
PyPDF2==3.0.1
python-docx==0.8.11
